requirements-test.txt
PyHamcrest==1.8.0
nose==1.3.0
nose-testconfig==0.9
requests==2.0.1
splinter==0.9.0
behave==1.2.6
